updated android NDK version

+++ src/libkubo/compile_android.sh
@@ -8,12 +8,29 @@ cd $SCRIPT_DIR
 set -e # Exit if any command fails
 set -x # Print commands for debugging
 
+# Assert Go version
+REQUIRED_GO_VERSION="1.19"
+INSTALLED_GO_VERSION=$(go version | awk '{print $3}' | sed 's/go//')
+
+# Compare versions using sort -V (version-aware)
+if [ "$INSTALLED_GO_VERSION" != "$REQUIRED_GO_VERSION" ]; then
+  echo "Error: Go $REQUIRED_GO_VERSION is required. Installed: $INSTALLED_GO_VERSION"
+  exit 1
+fi
+
 export GOOS=android
 export GOARCH=arm64
 export CGO_ENABLED=1
-export NDK_PATH=~/.buildozer/android/platform/android-ndk-r25b
+export NDK_PATH=~/.buildozer/android/platform/android-ndk-r28c
 export TOOLCHAIN=$NDK_PATH/toolchains/llvm/prebuilt/linux-x86_64
 export CC=$TOOLCHAIN/bin/aarch64-linux-android21-clang
+
+if ! [ -e $CC ];then
+  echo "The configured CC path doesn't exist:"
+  echo "$CC"
+  echo "Maybe you upgraded to a newer Android NDK version?"
+  exit 1
+fi
 
 # Clean old files
 rm -f ./libkubo_android_arm64.so ./libkubo_android_arm64.h
@@ -23,7 +40,7 @@ go mod tidy
 go build -v -buildmode=c-shared -o ./libkubo_android_arm64.so .
 
 echo "Build completed"
-ls -la ./libkubo_linux_x86_64.so
+ls -la ./libkubo_android_arm64.so
 
 exit 0
 """
